% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{add_button}
\alias{add_button}
\title{Add Buttons}
\usage{
add_button(answer, to, label)
}
\arguments{
\item{answer}{the answer to add the button to}

\item{to}{the answer to which the button should link}

\item{label}{the label on the button}
}
\value{
the answer with the added button
}
\description{
Add Buttons
}
