% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{answer}
\alias{answer}
\title{Bot Answer}
\usage{
answer(text)
}
\arguments{
\item{text}{the response text of the answer}
}
\value{
an answer object
}
\description{
Bot Answer
}
\details{
You can use basic Markdown syntax
}
