% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{answer_keyboard}
\alias{answer_keyboard}
\title{create keyboard from answer}
\usage{
answer_keyboard(answer)
}
\arguments{
\item{answer}{answer to generate keyboard from}
}
\value{
telegram.bot keyboard object
}
\description{
create keyboard from answer
}
