% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_telbot_interface.R
\name{answer_with}
\alias{answer_with}
\title{react to request for answers}
\usage{
answer_with(answer, bot, update)
}
\arguments{
\item{answer}{the answer}

\item{bot}{telegram.bot bot object}

\item{update}{telegram.bot update object}
}
\description{
react to request for answers
}
