% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_df.R
\name{batch_answers}
\alias{batch_answers}
\title{Create many answers at once}
\usage{
batch_answers(texts, names)
}
\arguments{
\item{texts}{vector of response texts}

\item{names}{names of the answers}
}
\value{
list of answers
}
\description{
Create many answers at once
}
