% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{command}
\alias{command}
\title{Answer to Command}
\usage{
command(command, answer)
}
\arguments{
\item{command}{character string with the name of the command (without the forward slash)}

\item{answer}{the answer object to be linked to the command}
}
\value{
the corresponding telegram.bot command handler
}
\description{
Answer to Command
}
