% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{get_answer_buttons}
\alias{get_answer_buttons}
\title{extract buttons from answer}
\usage{
get_answer_buttons(answer)
}
\arguments{
\item{answer}{answer object}
}
\value{
data frame of buttons needed for this answer
}
\description{
extract buttons from answer
}
