% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{has_buttons}
\alias{has_buttons}
\title{check if an answer has buttons attached}
\usage{
has_buttons(answer)
}
\arguments{
\item{answer}{an answer object}
}
\value{
logical TRUE if answer has buttons
}
\description{
check if an answer has buttons attached
}
