% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{is.answer}
\alias{is.answer}
\title{check if an object is a telegram bot "answer" object}
\usage{
is.answer(x)
}
\arguments{
\item{x}{object}
}
\value{
logical, TRUE if the class of `x` includes "telbot_answer"
}
\description{
check if an object is a telegram bot "answer" object
}
