% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_telbot_interface.R
\name{is_requested}
\alias{is_requested}
\title{check if answer is requested by telegram.bot update}
\usage{
is_requested(answer, update)
}
\arguments{
\item{answer}{answer to check}

\item{update}{telgram.bot update object}
}
\value{
logical TRUE if answer is requested
}
\description{
check if answer is requested by telegram.bot update
}
