% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_class.R
\name{print.telbot_answer}
\alias{print.telbot_answer}
\title{print telbot answers}
\usage{
\method{print}{telbot_answer}(x, ...)
}
\arguments{
\item{x}{an answer object}

\item{...}{ignored}
}
\description{
print telbot answers
}
