% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyboards.R
\name{quick_keyboard}
\alias{quick_keyboard}
\title{shorthand to create telegram.bot keyboard from key/value pairs}
\usage{
quick_keyboard(...)
}
\arguments{
\item{...}{argument name becomes data, value becomes label: quick_keyboard(data_send_back = "a button label")}
}
\value{
a telegram.bot inline keyboard object
}
\description{
shorthand to create telegram.bot keyboard from key/value pairs
}
