% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/answer_telbot_interface.R
\name{send_answer}
\alias{send_answer}
\title{send answer object}
\usage{
send_answer(answer, bot, update)
}
\arguments{
\item{answer}{answer to send}

\item{bot}{telegram.bot bot object}

\item{update}{telegram.bot update object}
}
\description{
send answer object
}
