% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_bot.R
\name{start_bot}
\alias{start_bot}
\title{start the telegram chatbot}
\usage{
start_bot(token, ...)
}
\arguments{
\item{token}{your telegram bot api token}

\item{...}{answers and commands}
}
\description{
start the telegram chatbot
}
